move _check_internal up a level
